+++ src/types.cpp
@@ -79,6 +79,21 @@ std::string typestr(const SizedType &type)
       res += ")";
       return res;
     }
+    case Type::record: {
+      std::string res = "record";
+      size_t n = type.GetFieldCount();
+
+      res += " { ";
+      for (size_t i = 0; i < n; ++i) {
+        res += "." + type.GetField(i).name;
+        res += " = ";
+        res += typestr(type.GetField(i).type);
+        if (i != n - 1)
+          res += ", ";
+      }
+      res += " }";
+      return res;
+    }
     case Type::kstack_t:
     case Type::ustack_t:
       return type.stack_type.name();
@@ -153,6 +168,23 @@ bool SizedType::IsCompatible(const SizedType &t) const
     return false;
   }
 
+  if (IsRecordTy() && t.IsRecordTy()) {
+    if (GetFieldCount() != t.GetFieldCount())
+      return false;
+
+    for (ssize_t i = 0; i < GetFieldCount(); i++) {
+      const auto &field_left = GetField(i);
+      if (!t.HasField(field_left.name)) {
+        return false;
+      }
+
+      const auto &field_right = t.GetField(field_left.name);
+
+      if (!field_left.type.IsCompatible(field_right.type))
+        return false;
+    }
+  }
+
   return true;
 }
 
@@ -197,6 +229,23 @@ std::strong_ordering SizedType::operator<=>(const SizedType &t) const
     return std::strong_ordering::equal;
   }
 
+  if (IsRecordTy()) {
+    if (auto cmp = GetFieldCount() <=> t.GetFieldCount(); cmp != 0)
+      return cmp;
+
+    for (ssize_t i = 0; i < GetFieldCount(); i++) {
+      const auto &field_left = GetField(i);
+      const auto &field_right = t.GetField(i);
+      if (auto cmp = field_left.name <=> field_right.name; cmp != 0)
+        return cmp;
+
+      if (auto cmp = field_left.type <=> field_right.type; cmp != 0)
+        return cmp;
+    }
+
+    return std::strong_ordering::equal;
+  }
+
   if (auto cmp = GetSize() <=> t.GetSize(); cmp != 0)
     return cmp;
 
@@ -214,7 +263,7 @@ bool SizedType::IsByteArray() const
 bool SizedType::IsAggregate() const
 {
   return IsArrayTy() || IsByteArray() || IsTupleTy() || IsCStructTy() ||
-         IsStack();
+         IsStack() || IsRecordTy();
 }
 
 bool SizedType::IsStack() const
@@ -267,6 +316,7 @@ std::string typestr(Type t)
     case Type::array:    return "array";    break;
     case Type::buffer:   return "buffer";   break;
     case Type::tuple:    return "tuple";    break;
+    case Type::record:    return "record";    break;
     case Type::timestamp:return "timestamp";break;
     case Type::mac_address: return "mac_address"; break;
     case Type::cgroup_path_t: return "cgroup_path_t"; break;
@@ -505,6 +555,13 @@ SizedType CreateTuple(std::shared_ptr<Struct> &&tuple)
   return s;
 }
 
+SizedType CreateRecord(std::shared_ptr<Struct> &&record)
+{
+  auto s = SizedType(Type::record, record->size);
+  s.inner_struct_ = std::move(record);
+  return s;
+}
+
 SizedType CreateMacAddress()
 {
   auto st = SizedType(Type::mac_address, 6);
@@ -529,31 +586,37 @@ bool SizedType::IsSigned() const
 
 std::vector<Field> &SizedType::GetFields() const
 {
-  assert(IsTupleTy() || IsCStructTy());
+  assert(IsTupleTy() || IsCStructTy() || IsRecordTy());
   return inner_struct()->fields;
 }
 
 Field &SizedType::GetField(ssize_t n) const
 {
-  assert(IsTupleTy() || IsCStructTy());
+  assert(IsTupleTy() || IsCStructTy() || IsRecordTy());
   if (n >= GetFieldCount())
     throw util::FatalUserException("Getfield(): out of bounds");
   return inner_struct()->fields[n];
 }
 
+size_t SizedType::GetFieldIdx(const std::string &name) const
+{
+  assert(IsRecordTy());
+  return inner_struct()->GetFieldIdx(name);
+}
+
 ssize_t SizedType::GetFieldCount() const
 {
-  assert(IsTupleTy() || IsCStructTy());
+  assert(IsTupleTy() || IsCStructTy() || IsRecordTy());
   return inner_struct()->fields.size();
 }
 
 void SizedType::DumpStructure(std::ostream &os)
 {
-  assert(IsTupleTy());
+  assert(IsTupleTy() || IsRecordTy());
   if (IsTupleTy())
     os << "tuple";
   else
-    os << "struct";
+    os << "record";
   inner_struct()->Dump(os);
 }
 
@@ -562,7 +625,7 @@ ssize_t SizedType::GetInTupleAlignment() const
   if (IsByteArray())
     return 1;
 
-  if (IsTupleTy() || IsCStructTy())
+  if (IsTupleTy() || IsCStructTy() || IsRecordTy())
     return inner_struct()->align;
 
   if (GetSize() <= 2)
@@ -577,19 +640,19 @@ ssize_t SizedType::GetInTupleAlignment() const
 
 bool SizedType::HasField(const std::string &name) const
 {
-  assert(IsCStructTy());
+  assert(IsCStructTy() || IsRecordTy());
   return inner_struct()->HasField(name);
 }
 
 const Field &SizedType::GetField(const std::string &name) const
 {
-  assert(IsCStructTy());
+  assert(IsCStructTy() || IsRecordTy());
   return inner_struct()->GetField(name);
 }
 
 std::shared_ptr<Struct> SizedType::inner_struct() const
 {
-  assert(IsCStructTy() || IsTupleTy());
+  assert(IsCStructTy() || IsTupleTy() || IsRecordTy());
   return std::visit(
       [](const auto &v) {
         if constexpr (std::is_same_v<std::decay_t<decltype(v)>,
@@ -604,7 +667,7 @@ std::shared_ptr<Struct> SizedType::inner_struct() const
 
 std::shared_ptr<const Struct> SizedType::GetStruct() const
 {
-  assert(IsCStructTy() || IsTupleTy());
+  assert(IsCStructTy() || IsTupleTy() || IsRecordTy());
   return inner_struct();
 }
 
@@ -634,6 +697,18 @@ bool SizedType::FitsInto(const SizedType &t) const
 
     for (ssize_t i = 0; i < GetFieldCount(); i++) {
       if (!GetField(i).type.FitsInto(t.GetField(i).type))
+        return false;
+    }
+    return true;
+  }
+
+  if (IsRecordTy()) {
+    if (GetFieldCount() != t.GetFieldCount())
+      return false;
+
+    for (ssize_t i = 0; i < GetFieldCount(); i++) {
+      const auto &field = GetField(i);
+      if (!field.type.FitsInto(t.GetField(field.name).type))
         return false;
     }
     return true;